requirements.txt
# Voice Clone Studio - Universal Requirements
# Supports Windows, Linux, and macOS with platform-specific packages

# Core TTS and audio processing dependencies (all platforms)
numpy==1.26.4
qwen-tts
audio-separator
librosa
torchaudio
soundfile
sox
einops
huggingface_hub
onnxruntime

# Web UI and utilities
gradio
gradio_filelister @ file:wheel/gradio_filelister-0.4.0-py3-none-any.whl
diffusers
markdown
psutil

# MMAudio Sound Effects dependencies (torchvision installed with PyTorch in setup script)
open_clip_torch
torchdiffeq
timm
colorlog
omegaconf
av>=14.0.1

# LuxTTS voice cloning (optional - install via setup script or manually)
# Prerequisites (install in order before zipvoice):
#   pip install git+https://github.com/ysharma3501/LinaCodec.git
#   pip install piper-phonemize --find-links https://k2-fsa.github.io/icefall/piper_phonemize.html
#   pip install "zipvoice @ git+https://github.com/ysharma3501/LuxTTS.git"

# 4-bit quantization support (all platforms)
bitsandbytes

# Voice isolation and denoising
deepfilternet

# Chatterbox Voice Conversion & TTS (vendored in modules/chatterbox/)
s3tokenizer
conformer

# Optional performance enhancements (install separately for best performance):
# - Flash Attention 2 (fast): pip install flash-attn --no-build-isolation

# Note for Linux users:
# - If onnxruntime fails to install, try the nightly build:
# - openai-whisper is excluded due to compatibility issues (llvmlite/numba build failures)
# - VibeVoice ASR can be used for transcription instead
